refactor(native): move the tool-use loop into its own module
`execute_with_code_intelligence` and the skill-tool dispatch it drives —
`skill_tool_outcome`, `skill_tool_provenance`, `set_skill_result_summary`,
`emit_skill_tool_lifecycle`, `AgentSkillToolLifecycle`, `dispatch_skill_tool` —
move to `api_process_adapter/execution.rs`.

`execute_with_code_intelligence` is moved, not decomposed. Its 978 lines are
byte-identical to their pre-split text apart from the `pub(super)` on the
declaration line. Splitting it means finding seams in live control flow that
handles tool dispatch, code-intelligence fallback, and error propagation; that
is its own change with its own review, not something to smuggle into a
relocation whose safety argument is "nothing changed but the file it lives in".

A `[ARCH-NATIVE-006]` entry records the module at 1,166 lines so the residual
is a tracked number rather than an unremarked one.
